+++ CHANGELOG.md
@@ -13,6 +13,9 @@
 - Add protections for `static/css/custom.css` and `static/js/custom.js`
 - Add `--fail-fast` CLI option
 - Add configuration options for `WriteReports`
+- Added syntax for parsing HTML comments in Markdown as attributes to apply to rendered HTML
+- Added syntax for parsing specially formatted text inside Markdown links, images, and fenced code blocks as attributes to apply to rendered HTML
+- Added documentation page with a table of supported Markdown features
 
 ### Fixed
 
@@ -40,8 +43,6 @@
 ### Removed
 
 - Removed Sitemap's stage configuration
-
-### Fixed
 
 ## [0.5.10] - 2026-08-02